producer_requirements.txt
# Kafka Producer Requirements
kafka-python==2.0.2
pandas==2.1.4
python-dateutil==2.8.2
